refactor(widgets): split widget-api into widget-model + widget-api
Layering fix. :widget-api previously mixed pure data shapes (ids,
LayoutGraph, @Widget annotation, DefaultLayout) with Compose-bound
surface (WidgetDescriptor, Registry, SlotRenderer, Locals). When
:client-launcher took a dep on :widget-api to host LayoutGraph
persistence, it transitively pulled compose-runtime -- the launcher
domain should be UI-framework-agnostic so a future TUI / CLI surface
can share persistence without touching Compose.

Two modules now:
- :widget-model -- pure JVM + kotlinx-serialization. Owns @Widget,
  SurfaceId, SlotId, WidgetKind, WidgetInstance, SlotContent,
  SurfaceLayout, LayoutGraph, DefaultLayout (with the bundled
  default-layout.json resource).
- :widget-api -- depends on :widget-model + compose-runtime via
  `api(project(":widget-model"))`. Owns WidgetDescriptor,
  WidgetRegistry, SlotRenderer, LocalLayoutGraph, LocalWidgetRegistry.

:client-launcher drops :widget-api, picks up :widget-model only --
no Compose on the launcher classpath. :client-ui keeps :widget-api
(transitively gets :widget-model). Widget processor scans
hivens.widget.model.Widget / WidgetInstance FQNs and generates a
registry that imports from both packages.
