+++ settings.gradle.kts
@@ -69,5 +69,6 @@ include(":client-config")
 include(":client-core")
 include(":client-launcher")
 include(":client-ui")
+include(":widget-model")
 include(":widget-api")
 include(":widget-processor")
