+++ widget-api/build.gradle.kts
@@ -1,20 +1,13 @@
 plugins {
     alias(libs.plugins.kotlin.jvm)
     alias(libs.plugins.kotlin.compose.compiler)
-    alias(libs.plugins.kotlin.serialization)
 }
 
+// Compose-bound surface for the widget kernel. Data shapes (ids,
+// layout graph, default-layout resource, @Widget annotation) live in
+// :widget-model so non-UI surfaces -- launcher persistence today,
+// future CLI / TUI -- can consume them without dragging compose-runtime.
 dependencies {
-    // @Composable annotation + Compose runtime types referenced by
-    // WidgetDescriptor.Render. compose-runtime is the smallest dep that
-    // brings in the annotation; the full compose plugin is overkill for
-    // a contracts module.
+    api(project(":widget-model"))
     implementation(libs.compose.runtime)
-    implementation(libs.kotlinx.serialization.json)
-
-    testImplementation(kotlin("test"))
-}
-
-tasks.test {
-    useJUnitPlatform()
 }
