+++ widget-model/src/main/kotlin/hivens/widget/model/DefaultLayout.kt
@@ -1,4 +1,4 @@
-package hivens.widget.api
+package hivens.widget.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -7,12 +7,12 @@ import kotlinx.serialization.json.Json
 private const val DEFAULT_LAYOUT_RESOURCE = "/widget/default-layout.json"
 
 // Reads the kernel-bundled default LayoutGraph out of the classpath.
-// The resource lives in :widget-api so the launcher does not need a
-// direct dependency on :client-ui to fetch it.
+// Resource ships inside :widget-model so the launcher (and any other
+// non-UI surface) can fetch it without depending on :client-ui.
 //
 // Failure modes (missing resource, malformed JSON) throw -- the
-// resource ships inside the jar and a failure means a broken build,
-// not a runtime user condition.
+// resource is part of the jar; a failure means a broken build, not
+// a runtime user condition.
 object DefaultLayout {
     fun load(json: Json = LENIENT): LayoutGraph {
         val stream = DefaultLayout::class.java.getResourceAsStream(DEFAULT_LAYOUT_RESOURCE)
