+++ widget-model/src/main/kotlin/hivens/widget/model/LayoutGraph.kt
@@ -1,4 +1,4 @@
-package hivens.widget.api
+package hivens.widget.model
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
