+++ widget-model/src/main/kotlin/hivens/widget/model/Widget.kt
@@ -1,4 +1,4 @@
-package hivens.widget.api
+package hivens.widget.model
 
 // Marker annotation scanned by the KSP processor in :widget-processor.
 // Apply to top-level @Composable functions with signature
@@ -8,6 +8,10 @@ package hivens.widget.api
 // id MUST be unique across the whole runtime. Convention:
 //   "<surface>.<role>"        for kernel widgets ("home.classic.header")
 //   "<plugin-id>.<role>"      for plugin-contributed widgets (Phase 3)
+//
+// Lives in :widget-model rather than :widget-api so it carries no
+// Compose dependency. The @Composable requirement is enforced by the
+// processor, not the annotation's classpath.
 @Target(AnnotationTarget.FUNCTION)
 @Retention(AnnotationRetention.BINARY)
 annotation class Widget(
