+++ widget-model/src/main/resources/widget/default-layout.json
